refactor: Modernize PHP SDK with enhanced type safety and new DTO models
New Features:
- Introduced comprehensive DTO models for campaigns, variations, targeting, and troubleshooting
- Added Types trait for centralized type definitions with PHPStan-level type safety
- Implemented ContainerException for better error handling
- Enhanced visitor strategies with context updates and troubleshooting support

Improvements:
- Removed deprecated FlagshipLogManager8 class
- Updated ModificationsDTO value type from mixed to array for consistency
- Refactored ContainerInterface with improved type specifications
- Updated HttpClient with better error handling and type safety
- Enhanced MurmurHash, Curl, and logging interfaces for clarity
- Added utility methods in Helper trait for array operations
- Improved flag activation with flag key inclusion in hits
- Updated return types across DecisionManagerAbstract and various interfaces

Testing:
- Refactored test suite to use BaseTestCase for consistency
- Added comprehensive unit tests for new DTO classes
- Enhanced test coverage for FlagshipLogManager, HttpClient, and strategy classes
- Improved test readability with standardized formatting and assertions

Configuration:
- Updated PHPStan configuration with new rules and future compatibility
- Enhanced composer.json with additional static analysis tools
- Refined phpstan.neon and added phpstan-future.neon

+++ src/Utils/HttpClientInterface.php
@@ -15,7 +15,7 @@ interface HttpClientInterface
     /**
      * Add extra headers to include in the request.
      *
-     * @param  array $headers : Collection key, value of http header
+     * @param  array<string, string> $headers : Collection key, value of http header
      * @return HttpClientInterface
      */
     public function setHeaders(array $headers): HttpClientInterface;
@@ -23,16 +23,16 @@ interface HttpClientInterface
     /**
      * set the Timeout
      *
-     * @param int $seconds
+     * @param float $seconds
      * @return HttpClientInterface
      */
-    public function setTimeout(int $seconds = FlagshipConstant::REQUEST_TIME_OUT): HttpClientInterface;
+    public function setTimeout(float $seconds = FlagshipConstant::REQUEST_TIME_OUT): HttpClientInterface;
 
     /**
      * send a http get request
      *
      * @param string $url
-     * @param  array $params Collection key, value of http params
+     * @param  array<mixed> $params Collection key, value of http params
      * @return HttpResponse
      */
     public function get(string $url, array $params = []): HttpResponse;
@@ -41,8 +41,8 @@ interface HttpClientInterface
      * send a http post request
      *
      * @param string $url
-     * @param  array $query Collection key, value of http params
-     * @param  array $data   Collection key, value of http post body
+     * @param  array<mixed> $query Collection key, value of http params
+     * @param  array<mixed> $data   Collection key, value of http post body
      * @return HttpResponse
      */
     public function post(string $url, array $query = [], array $data = []): HttpResponse;
